requirements.txt
elasticsearch 
transformers
datasets 
python-dotenv
torch==1.13